+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 Each entry (version) should have a date and one or more of 3 sections: 'Features', 'Bug Fixes', 'Breaking Changes'. Make sure to `git tag` the commit to match the version. Sub / pre-release versions should be hypenated (i.e. 1.0.3-2).
 
+# 0.0.3-1 (2015-04-01)
+
+## Features
+- accounts-password: add resentEnrollmentEmail method
+
+
+# 0.0.2 (2015-03-31)
+
+## Bug Fixes
+- password-reset: log out user first if logged in
+
+## Features
+- style: list class updates
+- accounts-password: return userId of created user
+- style: `.a-div` no underline, add `.list-even-odd`
+
+
 # 0.0.1 (2015-03-26)
+
 ## Features
 - [init, start changelog]